init obs before vue
