+++ app/app.ts
@@ -1,4 +1,4 @@
-import { I18nService } from 'services/i18n';
+import { I18nService, $t } from 'services/i18n';
 
 window['eval'] = global.eval = () => {
   throw new Error('window.eval() is disabled for security');
@@ -9,6 +9,7 @@ import Vue from 'vue';
 
 import { createStore } from './store';
 import { WindowsService } from './services/windows';
+import { ObsUserPluginsService } from 'services/obs-user-plugins';
 import { AppService } from './services/app';
 import Utils from './services/utils';
 import electron from 'electron';
@@ -24,6 +25,11 @@ import OneOffWindow from 'components/windows/OneOffWindow.vue';
 import electronLog from 'electron-log';
 import { UserService, setSentryContext } from 'services/user';
 import { getResource } from 'services';
+import * as obs from '../obs-api';
+import path from 'path';
+import uuid from 'uuid/v4';
+
+const crashHandler = window['require']('crash-handler');
 
 const { ipcRenderer, remote } = electron;
 const slobsVersion = remote.process.env.SLOBS_VERSION;
@@ -119,11 +125,70 @@ document.addEventListener('dragover', event => event.preventDefault());
 document.addEventListener('dragenter', event => event.preventDefault());
 document.addEventListener('drop', event => event.preventDefault());
 
+export const apiInitErrorResultToMessage = (resultCode: obs.EVideoCodes) => {
+  switch (resultCode) {
+    case obs.EVideoCodes.NotSupported: {
+      return $t('OBSInit.NotSupportedError');
+    }
+    case obs.EVideoCodes.ModuleNotFound: {
+      return $t('OBSInit.ModuleNotFoundError');
+    }
+    default: {
+      return $t('OBSInit.UnknownError');
+    }
+  }
+};
+
+const showDialog = (message: string): void => {
+  electron.remote.dialog.showErrorBox($t('OBSInit.ErrorTitle'), message);
+};
+
 document.addEventListener('DOMContentLoaded', () => {
   createStore().then(async store => {
-    // handle closeWindow event from the main process
     const windowsService: WindowsService = WindowsService.instance;
+
     if (Utils.isMainWindow()) {
+      // Services
+      const appService: AppService = AppService.instance;
+      const obsUserPluginsService: ObsUserPluginsService = ObsUserPluginsService.instance;
+
+      // This is used for debugging
+      window['obs'] = obs;
+
+      // Host a new OBS server instance
+      obs.IPC.host(`slobs-${uuid()}`);
+      obs.NodeObs.SetWorkingDirectory(
+        path.join(
+          electron.remote.app.getAppPath().replace('app.asar', 'app.asar.unpacked'),
+          'node_modules',
+          'obs-studio-node',
+        ),
+      );
+
+      crashHandler.registerProcess(appService.pid, false);
+
+      await obsUserPluginsService.initialize();
+
+      // Initialize OBS API
+      const apiResult = obs.NodeObs.OBS_API_initAPI(
+        'en-US',
+        appService.appDataDirectory,
+        electron.remote.process.env.SLOBS_VERSION,
+      );
+
+      if (apiResult !== obs.EVideoCodes.Success) {
+        const message = apiInitErrorResultToMessage(apiResult);
+        showDialog(message);
+
+        crashHandler.unregisterProcess(appService.pid);
+
+        obs.NodeObs.StopCrashHandler();
+        obs.IPC.disconnect();
+
+        electron.ipcRenderer.send('shutdownComplete');
+        return;
+      }
+
       ipcRenderer.on('closeWindow', () => windowsService.closeMainWindow());
       AppService.instance.load();
     } else {
