+++ app/services/obs-user-plugins.ts
@@ -4,6 +4,9 @@ import path from 'path';
 import electron from 'electron';
 import * as obs from '../../obs-api';
 
+// WARNING: This service is initialized extremely early
+// and should not import any other services.
+
 export class ObsUserPluginsService extends Service {
   async initialize() {
     // Make a best effort but don't stop SLOBS from loading
