+++ app/services/user.ts
@@ -127,7 +127,6 @@ export class UserService extends PersistentStatefulService<IUserServiceState> {
   // Makes sure the user's login is still good
   validateLogin() {
     if (!this.isLoggedIn()) return;
-    console.log('makes sure the login');
 
     const host = this.hostsService.streamlabs;
     const headers = authorizedHeaders(this.apiToken);
